Try before_save callback. Seems to have the exact same (functional) results as before_create
which is not what I was expecting. The instructions suggest I would get an error, or something would be weird, but the password is stored and saved (at least according to the puts verifaction line I've included in ex8.

I would go for after_create, simply because the password is not then printed to the console (unless I explicitly set it to, with the puts that I'm using). But some dude on SO suggests these are all bad approaches so, I don't know.